+++ rasa_sdk/__main__.py
@@ -3,6 +3,7 @@ import logging
 from rasa_sdk import utils
 from rasa_sdk.endpoint import create_argument_parser, run
 from rasa_sdk.constants import APPLICATION_ROOT_LOGGER_NAME
+from rasa_sdk.tracing.utils import get_tracer_provider
 
 
 def main_from_args(args):
@@ -17,7 +18,7 @@ def main_from_args(args):
         args.logging_config_file,
     )
     utils.update_sanic_log_level()
-    tracer_provider = utils.get_tracer_provider(args)
+    tracer_provider = get_tracer_provider(args)
 
     run(
         args.actions,
